add route plugin, register route in index.ts
